Split out the archive and upload into a base class that does not include reports or schedules
