+++ BridgeApp/BridgeApp/Study Management/SBAScheduleManager.swift
@@ -52,7 +52,7 @@ extension Notification.Name {
 /// activities, but will *not* cache them all in memory. Instead, it will filter out those activites that are
 /// valid for today and the most recent finished activity (if any) for each activity identifier where the
 /// "activity identifier" refers to an `SBBActivity` object's associated `SBAActivityReference`.
-open class SBAScheduleManager: SBAReportManager, RSDDataArchiveManager {
+open class SBAScheduleManager: SBAReportManager {
 
     /// List of keys used in the notifications sent by this manager.
     public enum NotificationKey : String {
@@ -202,9 +202,6 @@ open class SBAScheduleManager: SBAReportManager, RSDDataArchiveManager {
     }
     private var _isReloadingScheduledActivities: Bool = false
     
-    /// A serial queue used to manage data crunching.
-    public let offMainQueue = DispatchQueue(label: "org.sagebionetworks.BridgeApp.SBAScheduleManager")
-    
     /// Reload the data by fetching changes to the scheduled activities.
     override open func reloadData() {
         loadScheduledActivities()
@@ -337,7 +334,7 @@ open class SBAScheduleManager: SBAReportManager, RSDDataArchiveManager {
     ///     - taskResult: The task result for this task, subtask, or section.
     ///     - scheduleIdentifier: The schedule identifier that was used to start the task.
     /// - returns: The schedule associated with this task view controller (if found).
-    open func scheduledActivity(for taskResult: RSDTaskResult, scheduleIdentifier: String?) -> SBBScheduledActivity? {
+    override open func scheduledActivity(for taskResult: RSDTaskResult, scheduleIdentifier: String?) -> SBBScheduledActivity? {
         
         let todayPredicate = SBBScheduledActivity.availableTodayPredicate()
         
@@ -666,132 +663,4 @@ open class SBAScheduleManager: SBAReportManager, RSDDataArchiveManager {
                                             userInfo: [NotificationKey.updateScheduleGuids : guids])
         }
     }
-    
-    /// Archive and upload results.
-    ///
-    /// DO NOT MAKE OPEN. This method retains the task path until archiving is completed and because it
-    /// nils out the pointer to the task path with a strong reference to `self`, it will also retain the
-    /// schedule manager until the completion block is called. syoung 05/31/2018
-    private final func archiveAndUpload(taskViewModel: RSDTaskViewModel) {
-        let uuid = UUID()
-        self._retainedPaths[uuid] = taskViewModel
-        taskViewModel.archiveResults(with: self) {
-            self._retainedPaths[uuid] = nil
-        }
-    }
-    private var _retainedPaths: [UUID : RSDTaskViewModel] = [:]
-    
-    // MARK: RSDDataArchiveManager
-    
-    /// Should the task result archiving be continued if there was an error adding data to the current
-    /// archive? Default behavior is to flush the archive and then return `false`.
-    ///
-    /// - parameters:
-    ///     - archive: The current archive being built.
-    ///     - error: The encoding error that was thrown.
-    /// - returns: Whether or not archiving should continue. Default = `false`.
-    open func shouldContinueOnFail(for archive: RSDDataArchive, error: Error) -> Bool {
-        debugPrint("ERROR! Failed to archive results: \(error)")
-        // Flush the archive.
-        (archive as? SBBDataArchive)?.remove()
-        return false
-    }
-    
-    /// When archiving a task result, it is possible that the results of a task need to be split into
-    /// multiple archives -- for example, when combining two or more activities within the same task. If the
-    /// task result components should be added to the current archive, then the manager should return
-    /// `currentArchive` as the response. If the task result *for this section* should be ignored, then the
-    /// manager should return `nil`. This allows the application to only upload data that is needed by the
-    /// study, and not include information that is ignored by *this* study, but may be of interest to other
-    /// researchers using the same task protocol.
-    open func dataArchiver(for taskResult: RSDTaskResult, scheduleIdentifier: String?, currentArchive: RSDDataArchive?) -> RSDDataArchive? {
-
-        // Look for a schema info associated with this portion of the task result. If not found, then
-        // return the current archive.
-        let schema = taskResult.schemaInfo ?? self.schemaInfo(for: taskResult.identifier)
-        guard (currentArchive == nil) || (schema != nil) else {
-            return currentArchive
-        }
-        
-        let schemaInfo = schema ?? RSDSchemaInfoObject(identifier: taskResult.identifier, revision: 1)
-        let archiveIdentifier = schemaInfo.schemaIdentifier ?? taskResult.identifier
-        let schedule = self.scheduledActivity(for: taskResult, scheduleIdentifier: scheduleIdentifier)
-            ?? (currentArchive as? SBAScheduledActivityArchive)?.schedule
-        let isPlaceholder = (currentArchive == nil) && (schema == nil) && (schedule == nil)
-        
-        // If there is a top-level archive then return the exisiting if and only if the identifiers are the
-        // same or the schema is nil.
-        if let inputArchive = currentArchive,
-            ((inputArchive.identifier == archiveIdentifier) || (schema == nil)) {
-            return inputArchive
-        }
-        
-        // Otherwise, instantiate a new archive.
-        return SBAScheduledActivityArchive(identifier: archiveIdentifier, schemaInfo: schemaInfo, schedule: schedule, isPlaceholder: isPlaceholder)
-    }
-    
-    /// Finalize the upload of all the created archives.
-    public final func encryptAndUpload(taskViewModel: RSDTaskViewModel, dataArchives: [RSDDataArchive], completion:@escaping (() -> Void)) {
-        let archives: [SBBDataArchive] = dataArchives.compactMap {
-            guard let archive = $0 as? SBBDataArchive, self.shouldUpload(archive: archive) else { return nil }
-            return archive
-        }
-        #if DEBUG
-        archives.forEach {
-            guard let archive = $0 as? SBAScheduledActivityArchive else { return }
-            self.copyTestArchive(archive: archive)
-        }
-        #endif
-        SBBDataArchive.encryptAndUploadArchives(archives)
-        completion()
-    }
-    
-    /// This method is called during `encryptAndUpload()` to allow subclasses to cancel uploading an archive.
-    ///
-    /// - returns: Whether or not to upload. Default is to return `true` if the archive is not empty.
-    open func shouldUpload(archive: SBBDataArchive) -> Bool {
-        return !archive.isEmpty()
-    }
-    
-    /// By default, if an archive fails, the error is printed and that's all that is done.
-    open func handleArchiveFailure(taskViewModel: RSDTaskViewModel, error: Error, completion:@escaping (() -> Void)) {
-        debugPrint("WARNING! Failed to archive \(taskViewModel.identifier). \(error)")
-        completion()
-    }
-    
-    #if DEBUG
-    private func copyTestArchive(archive: SBAScheduledActivityArchive) {
-        guard SBAParticipantManager.shared.isTestUser else { return }
-        do {
-            if !archive.isCompleted {
-                try archive.complete()
-            }
-            let fileManager = FileManager.default
-            
-            let outputDirectory = try fileManager.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: true)
-            let dirURL = outputDirectory.appendingPathComponent("archives", isDirectory: true)
-            try FileManager.default.createDirectory(at: dirURL, withIntermediateDirectories: true, attributes: nil)
-            
-            // Scrub non-alphanumeric characters from the identifer
-            var characterSet = CharacterSet.alphanumerics
-            characterSet.invert()
-            var filename = archive.identifier
-            while let range = filename.rangeOfCharacter(from: characterSet) {
-                filename.removeSubrange(range)
-            }
-            filename.append("-")
-            let dateFormatter = DateFormatter()
-            dateFormatter.locale = Locale(identifier: "en_US_POSIX")
-            dateFormatter.dateFormat = "yyyy-MM-dd'T'HHmm"
-            let dateString = dateFormatter.string(from: Date())
-            filename.append(dateString)
-            let debugURL = dirURL.appendingPathComponent(filename, isDirectory: false).appendingPathExtension("zip")
-            try fileManager.copyItem(at: archive.unencryptedURL, to: debugURL)
-            debugPrint("Copied archive to \(debugURL)")
-            
-        } catch let err {
-            debugPrint("Failed to copy archive: \(err)")
-        }
-    }
-    #endif
 }
